throw when initing existing file

+++ src/potent/commands/init.py
@@ -14,17 +14,19 @@ def init(
     path: Annotated[
         Path,
         typer.Argument(
-            exists=False,
             dir_okay=False,
             resolve_path=True,
             callback=is_plan_json,
-            help="The location in which to write a new `.plan.json` file",
+            help="The location in which to write a new `.plan.json` file. Must not exist",
         ),
     ],
 ):
     """
     Create an empty plan at the specified path.
     """
+    if path.exists():
+        raise ValueError(f"A file already exists at {path}")
+
     path.write_text(
         Plan(version="v1", steps=[], directories=[]).model_dump_json(indent=2)
     )
